simplified the center piece
The spool and the center piece are now a single item which means that we
don't need a screw and a screwhole that might be difficult to 3D print.
On the other hand, there are a few problematic overhangs in the object
so it might not be that easy to 3D print. Let's test which approach
works better.